+++ plutarch-test/tests/Plutarch/Extra/MaybeSpec.hs
@@ -1,13 +1,12 @@
 module Plutarch.Extra.MaybeSpec (spec) where
 
-import Plutarch.Extra.Maybe
-import Plutarch.Prelude
-
 import Data.Maybe (fromJust, isJust)
 import Hedgehog (Property)
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Internal.Property (propertyTest)
-import Plutarch.Api.V1.Maybe (PMaybeData)
+import Plutarch.Api (PMaybeData)
+import Plutarch.Extra.Maybe
+import Plutarch.Prelude
 import Plutarch.Test
 import Plutarch.Test.Property
 import Plutarch.Test.Property.Gen